uui | fix: generate is deprecated
