+++ testing/main.go
@@ -1,4 +1,4 @@
 package ldetesting
 
-//go:generate ldetool generate --little-endian --package ldetesting parsing.lde
-//go:generate ldetool generate --package ldetesting --go-string string.lde
+//go:generate ldetool --little-endian --package ldetesting parsing.lde
+//go:generate ldetool --package ldetesting --go-string string.lde
